Time: 4 ms (82.77%) | Memory: 42 MB (85.07%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean judgeSquareSum(int c) {
         long left = 0;
-        long right = (long)Math.sqrt(c);
+        long right = (int)Math.sqrt(c);
 
         while (left <= right) {
             long sum = left * left + right * right;
